+++ Runtime/Tracker/ReleaseDropOff.cs
@@ -1,0 +1,26 @@
+namespace LiteNinja.MusicForge
+{
+  /// <summary>
+  /// Enumeration of different release drop off curves.
+  /// </summary>
+  public enum ReleaseDropOff
+  {
+    /// <summary>
+    /// Linear release drop off curve
+    /// </summary>
+    Linear,
+    /// <summary>
+    /// Shallow release drop off curve
+    /// </summary>
+    ShallowCurve,
+    /// <summary>
+    /// Medium release drop off curve
+    /// </summary>
+    MediumCurve,
+    /// <summary>
+    /// Steep release drop off curve
+    /// </summary>
+    SteepCurve
+  }
+
+}